testing combbo wrapping

+++ header/simulator.cpp
@@ -4,10 +4,10 @@
 
 #include "simulator.h"
 
-int sampling(std::vector<double> *probabilities) {
+int sampling(std::vector<double>* probabilities) {
 
     double total = 0;
-    for (int i = 0; i < probabilities->size(); ++i) total = total + probabilities->at(i);
+    for (size_t i = 0; i < probabilities->size(); ++i) total = total + probabilities->at(i);
     if (total != 1.) throw std::invalid_argument("\n\nProbabilities don't sum to 1");
 
     std::discrete_distribution<int> dist(probabilities->begin(), probabilities->end());
